New query infrastructure; generalised backmatter computed from queries
1. Improved querying

This patch improves the querying infrastucture to be strong enough to
compute the existing backmatter, as well as a few other things (like
computing the hereditary contributors to a given tree).

One breaking change is that I now want to support only "graph-related"
queries: previously, there was `\query/meta{key}{body}`, which would do
a structural equality match on the provided Forester content `body`. I
think queries like this are brittle and not too useful, and should be
replaced (in the future) with usage of the impending custom graph
support. Ideally, most queries should be able to start by reading a
small fraction of the forest from the graph and intersecting it with
things --- rather than walking the entire forest again and again.

2. Generalised backmatter

The backmatter is now generalised to support arbitrary queries (but this
doesn't yet have a user interface).This impacts the XML format of the
backmatter: previously it was something like:

    <backmatter>
     <context> ... </context>
     <references> ...  </references>
     <related> ...  </related>
     <backlinks> ... </backlinks>
    </backmatter>

Now, the backmatter element contains arbitrary trees. For example:

    <backmatter>
     <tree ...>
      <frontmatter>
       <title>Context</title>
       ...
      </frontmatter>
      ...
     </tree>
     ...
    </backmatter>

3. Miscellaneous

I have also disentangled and simplified much of the architecture. Here
are some of the additional changes.

+ The graph analysis is computed on the fly during evaluation rather
  than in a second pass.

+ The Render_effect thing is excised, and appropriate parameters are
  passed to modules where needed. Resource management (caches, etc.) is
  simplified across the board using generative module functors.

+ Previously the Xml_tree representation held routes for internal links,
  etc. I have removed this data from the representation because this is
  meant to be a source of truth for plenty of output formats, not all of
  which will route trees to the same locations as the built-in XML output
  format. Therefore, routing should in fact be computed during
  serialisation.

Closes: https://todo.sr.ht/~jonsterling/forester/63