added infrastructure for static build
Hi all,

This patch adds a docker file that builds a statically-linked forester
binary. We use this binary in github actions: see
https://github.com/ToposInstitute/CatColab/blob/main/dev-docs/ci-build.sh.

As much as I like nix, it's much speedier to just download a 4MB tarball
than it is to do the nix evaluation and pull forester in from cache.

I tried to get static building working through nix (it's supposedly
supported by opam-nix), but I was unfortunatly not able to get it
working; if someone else can get it working that would be great!

Also, I don't know how CI/releases on sourcehut work, but it would be
cool to host these statically linked binaries on some sort of official
release page; that would make it super easy to install forester!

Cheers,
-Owen