Switch to straightline latex-to-svg string pipeline
This compiles LaTeX to DVI in a temporary directory, and then pipes the
results through `dvisvgm` through stdin/stdout. The temporary directory
is deleted unless there was an error building the DVI file, in which
case the directory is left in place so that a user can diagnose the
problem.

Also adds error handling and diagnostic logging via Asai.

This is a first step to embedding the svgs directly in the tree file.

References: https://todo.sr.ht/~jonsterling/forester/82