Switch back from dvilualatex to latex
This is unfortunate, but necessary for now. With dvilualatex, users have
reported that the bounding boxes for tikzcd diagrams are computed
incorrectly, resulting in strange-looking margins.

I traced this to the fact that standalone.cls does not correctly set the
'papersize' special when executed under LuaLaTeX, rendering dvisvgm
--bbox=papersize ineffective. This problem could be solved using a
patched version of standalone.cls, which we should pursue through usual
channels, but for now, we don't really have a choice but to revert back
to LaTeX.

A future solution to this problem could be allowing user configuration
of external build jobs, but the programs that a user might want to run
(including LaTeX, etc.) don't exactly have a uniform interface: some
programs (like dvisvgm) can be run as a pipe, but others (like LaTeX)
cannot and must have a temporary directory. So a general
user-configurable solution to this problem is not exactly forthcoming
without introducing a ton of complexity.

References: https://todo.sr.ht/~jonsterling/forester/82