Add DSL for writing forester IR
The most interesting thing here is the xml_elt function, as it actually
uses the lexer. This is also the reason for moving the `split_xml_qname`
function to Forester_xml_names.

The motivation for this change is that it eases the development of the
hypermedia server. There is also a half-baked idea of a forester
vdom...