first cut at new version of publishing
You can do something like this:

  \publish-query{test-pub}\datalog{
    ?X -: {\rel/has-tag ?Z '{blog-post}} {\rel/in-bundle-closure ?Z ?X}
  }

This will then emit a json blob named 'publications/test-pub.json'
containing all the resources in the "bundle closure" of the set of blog
posts. The "bundle closure" is closed under both transclusion and
asset-routing. This means that if you link to an asset in a bundled tree
or use it in an <html:img> tag, it will get included in the bundle. I am
not totally sure if that behaviour is desirable, but we currently would
have no other way to get an asset to be included.